requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP Client
httpx>=0.25.0

# Vector Store
chromadb>=0.4.18

# Graph Database
nebula3-python>=3.5.0

# Utilities
python-dotenv>=1.0.0

# Optional: For production deployment
gunicorn>=21.2.0
